Fixed bug with parsing BigInteger in PUT method in backend; dates in CampaignPresenter.jsx now shows; added modal window close button; records statistic on download button

+++ course-04.09.20/backend/src/models/Campaign.kt
@@ -69,8 +69,8 @@ fun Route.campaign() {
             CampaignTable.update({ CampaignTable.camp_id eq id }) { c ->
                 f["name"]?.let { c[name] = it as String }
                 f["customer"]?.let { c[customer] = it as String }
-                f["earning"]?.let { c[earning] = it as BigDecimal }
-                f["spending"]?.let { c[spending] = it as BigDecimal }
+                f["earning"]?.let { c[earning] = (it as Double).toBigDecimal() }
+                f["spending"]?.let { c[spending] = (it as Double).toBigDecimal() }
                 f["executionStatus"]?.let { c[execution_status] = it as String }
             }
             "$raw updated)" to HttpStatusCode.OK
